refractured almost every component to make our can-of-books-frontend more organized
